core: extract device manager to separate file
